Make FileData null terminated and support mutable access.
I'd much prefer not to do this, but nanosvg's parse-from-memory function
(nsvgParse) takes a mutable null terminated string as input. So to use
that I need to do one of:
1 - Use nsvgParseFromFile and bypass the FileSystem layer
    (which means mods won't work). This is the current state.
2 - Make FileData null terminated and give a mutable access point.
3 - Take the FileData (which itself is a malloc allocated array in all the
    existing FileSystem implementations in the code base), and copy the
    data into a second array (or std::string or whatever) just so I can
    get mutable access and a null terminator. That just seems wasteful.

So instead, let's just botch the FileData interface (which is easy given
the current malloc based implementation) so it can be used directly.